% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaws-package.R
\docType{package}
\name{inlaws-package}
\alias{inlaws}
\alias{inlaws-package}
\title{Additional distribution families for mgcv}
\description{
Experimental response families for models fitted with \code{\link[mgcv:gam]{mgcv::gam()}}.
See each family for its supported fitting methods and parameterization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gam-mafia.github.io/inlaws/}
  \item \url{https://github.com/gam-mafia/inlaws}
  \item Report bugs at \url{https://github.com/gam-mafia/inlaws/issues}
}

}
\author{
\strong{Maintainer}: Gavin Simpson \email{ucfagls@gmail.com}

Authors:
\itemize{
  \item Gavin Simpson \email{ucfagls@gmail.com}
}

}
\keyword{internal}
